Refactor plugins/__init__.py for maintainability
