+++ linodecli/arg_helpers.py
@@ -177,7 +177,7 @@ def register_plugin(module, config, ops):
         msg = "Plugin name conflicts with CLI operation - registration failed."
         return msg, 12
 
-    if plugin_name in plugins.available_local:
+    if plugin_name in plugins.AVAILABLE_LOCAL:
         msg = "Plugin name conflicts with internal CLI plugin - registration failed."
         return msg, 13
 
@@ -219,7 +219,7 @@ def remove_plugin(plugin_name, config):
     """
     Remove a plugin
     """
-    if plugin_name in plugins.available_local:
+    if plugin_name in plugins.AVAILABLE_LOCAL:
         msg = f"{plugin_name} is bundled with the CLI and cannot be removed"
         return msg, 13
